Decompile classes related to dynamic linking (#106)
* Prepare headers for DMC

* Match and link `d_dylink` and `c_dylink`

* Mark unofficial symbols

* Remove symbols no longer needed

* Fix `.dtors` section in `d_profileNP.json`

* Apply code review (see description)
- Add documentation for the new classes
- Add stripped logic for dDyl::LinkAsync and dDyl::Unlink
- Add structure for nw4r::db::MapFile and constructor for DbMapFile
- Use ctor initializers where possible
- Various cleanup

---------

+++ include/lib/egg/core/eggHeap.h
@@ -28,7 +28,7 @@ public:
     virtual void *alloc(size_t, long) = 0;
     virtual void free(void *) = 0;
     virtual void destroy() = 0;
-    virtual size_t resizeForMBlock(void* block, size_t size) = 0;
+    virtual int resizeForMBlock(void* block, size_t size) = 0;
     virtual u32 getTotalFreeSize();
     virtual size_t getAllocatableSize(long align) = 0;
     virtual size_t adjust() = 0;
